requirements.txt
pygame-ce==2.5.0
black==24.4.2
pytmx==3.32